// src/Application/types.ts
export type Theme = {
  dark: boolean;
  colors: {
    primary: {
      default: string;
      light: string;
      container: string;
    };
    secondary: {
      default: string;
      light: string;
      container: string;
    };
    background: {
      default: string;
      surface: string;
      disable: string;
    };
    text: {
      default: string;
      secondary: string;
      hint: string;
      disable: string;
    };
    border: {
      default: string;
      disable: string;
    };
    success: {
      default: string;
      light: string;
      container: string;
    };
    warning: {
      default: string;
      light: string;
      container: string;
    };
    error: {
      default: string;
      light: string;
      container: string;
    };
  };
  font: string;
  assets?: {
    headerBackground?: ImageSourcePropType;
  };
};

/**
 * Context type.
 */
export type Context = {
  theme: Theme;
  navigator?: Navigator;
  translate: (key: string) => string;
};